Rename README.md to QA Reply.md
Hi,
 I will be writing the test cases below for Zomato Web App, Zomato Mobile app and Zomato API

Test Cases for Zomato Web App:
1.Verifying that Zomato app is opening in various browsers.(eg-Chrome,Firefox,IE)
2.Verifying that Zomato app works in various OS.eg-Windows,Mac,etc
3.Verifying that ZOmato app is not causing a problem for other app present in the system/machine already.
The above 3 test cases will form the base for Pre-conditions testing.

4.Verifying that functionality of the entire Zomato web app.
-This will include testing like clicks on all buttons/options
-Testing the GUI part like the placement of images and button and how they look on the web page.
-Checking for Inputs like whether user is able to type in the text box-eg Searching for a restaurant and user is able to type in the test box.
-Checking for the output is as per the input given-eg-Restaurant details for restaurant searched.
-Checking that proper value is returned from database as per the search done or query made.
 Based on Test Plan above few Test cases are as below:
a.Verify that user is able to select a city
b.Verify that user is able to type restaurant name in "Search for restaurant field".
c.Verify that user can give special characters or numbers or alphabets in "Search for Restaurant field"
d.Verify that result returned after search is relevant.
e.Verify on giving improper restaurant name, user gets proper information message stating the reason.
f.Verify that on selecting a city, restaurants of that city are only shown and not of other cities.eg City is Bangalore and restaurants are of Chennai. 
g.Verify that on clicking on "Search for restaurant", user gets option to select from a  list of options or  enter the restaurant name.
h.Verify that all information displayed for a  search is up-to-date eg-Location,menu,Timings,Pricing,etc

5. We also check for Non-functional part of the App which includes:
a.Security of the app - eg-If any payment is to be done
b.Recovery for the app - If the app crashes while doing payment or due to any other reason.
c.App Performance - Number of people visiting the app at same time so the load and stress testing comes in picture
d.Configuration,Compatibility testing - This includes different softwares/OS/Browsers/Hardware related.
e.Usability testing like how it looks,flow,reporting format,usage.

Once all these are done with positive testing and negative testing,we can move ahead with the release of the app.

6.Then it will include Smoke testing before the release

The same will go with Mobile App and with Zomato API.
Zomato API testing will be bit different as it only needs to be integrated and tested that it will have no effect on existing system. 
Also we can automated things for the APP as we move ahead.


Example for a Test Plan and the cases included for Mobile Zomato App:
ZOmato Mobile App"

Test plan for Download of app:
1.Verify user is able to download the app from Play Store.
2.Verify that all the permissions for data collection are shown to user before download begins.
3.Verify that user is able to open app on successful download.
4.Verify that if download halts in middle due to connectivity issue, it resumes are the issue is solved.

Test Plan for Functionality of app:
1.Verify user is able to scroll through the app smoothly.
2.Verify that user is able to select city, and tap on other options provided
3.Verify that user is able to tap on search and he gets a list of restaurant type to select or type restaurant name.
4.Verify that on selecting a particular type, those restaurant list are shown.
5.Verify that user can sort as per the option provided.
6.Verify that on selecting a restaurant, user is shown details about that restaurant only
7.Verify the GUI part of that page.
8.Verify for options on that page.


There are many  more cases and scenarios that can be taken up like app performance and wifi issue and many more. I will stop here for now.

For low literate users, the basic thing is language.
If user is know to the language, he is definitely going to use the app.
The UI and UX can be kept as simple as possible at the start of the page. The basic functionality of zomato is searching a restaurant so "Search" should be a most visible part on UI side.

For connectivity, taking India as an example,
Here majority people have Smartphone but the Internet usage in these smartphone is mostly 2G which is a slow. So keeping this in mind, the app should be designed in such a way that it takes least time to load the app and dividing the app city wise,makes it more better for searching restaurant  in app as it will filter out other restaurants giving a list of only that city.